[http client] Generation for @Client.Import when argument names lost
When the interface has been previously compiled then the method arg
names are effectively lost - become arg0, arg1, arg2 etc. In this case
the generation can get the implied arg names for path parameters from
the segments but for @QueryParam and @Header parameters we now need
explicitly named arguments.

This change handles the path segments and gives a reasonable error
for QueryParams and Headers etc that are not explicitly named.