Add Bearer authentication intercept
Currently, the documentation provides instructions on how to build a
bearer token provider, which assumes that tokens are short-lived and
must be refreshed after some time. However, some REST APIs offer a
bearer token authentication mechanism that relies on the API key/secret
instead of short-lived tokens. Since this is a somewhat common use case,
I feel like it makes sense to have this implementation offered out of
the box.